Moved stuff into internal, giving us better control
